requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.10.1
boto3==1.24.11
botocore==1.27.96
bs4==0.0.1
cachetools==4.2.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.2.0
distro==1.8.0
Django==3.2
django-cors-headers==3.11.0
django-environ==0.8.1
django-extensions==3.2.3
django-filter==21.1
django-redis==5.3.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.21.5
et-xmlfile==1.1.0
exceptiongroup==1.1.3
facebook-sdk==3.1.0
geographiclib==2.0
geopy==2.4.0
google==3.0.0
google-auth==1.21.2
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.3
MarkupSafe==2.1.3
mypy-extensions==1.0.0
numpy
openai==0.28
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.11.0
psycopg2-binary==2.9.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.5.0
pydantic_core==2.14.1
PyJWT==2.8.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.0
requests==2.31.0
rsa==4.9
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
s3transfer==0.6.2
selenium==4.13.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
wsproto==1.2.0
